src/functions/get_cost_report/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
sendgrid==6.*
